pyproject.toml
[project]
name = "insightchat"
version = "0.1.0"
description = "AI-powered chat application with RAG (Retrieval-Augmented Generation) support using Ollama"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=7.0.0",
]
